# performance improvement for time format