### Description
use custom format code instead of standard time.Time.Format

```text
BenchmarkGetTimeDateClockIndependent-8       10504020           114 ns/op           0 B/op           0 allocs/op
BenchmarkGetTimeDateClockTogether-8          23327814          48.1 ns/op           0 B/op           0 allocs/op

BenchmarkFormatDatetime-8                       12794754            72.7 ns/op         0 B/op           0 allocs/op
BenchmarkFormatDatetimeViaStandardFormat-8        4033940           303 ns/op          32 B/op          1 allocs/op
```
### Checklist
- [x] Code compiles correctly
- [ ] Created tests which fail without the change (if possible)
- [ ] All tests passing
- [ ] Extended the README / documentation, if necessary
- [x] Added myself / the copyright holder to the AUTHORS file