@methane 2020-05-30
Please include macro benchmark (benchmark through executing SQL) too for both of setting `interpolateParams` false and true.
Without it, we can't know this is real improvement.
Especially, this PR seems to increase allocation of interpolateParams.