@chanxuehong 2020-05-31
> Please include macro benchmark (benchmark through executing SQL) too for both of setting `interpolateParams` false and true.
> Without it, we can't know this is real improvement.
> Especially, this PR seems to increase allocation of interpolateParams.


Sorry, I do n’t know how to run two versions of the mysql driver in one test at the same time, so I did separate tests. The following is my tests:

here is the table
```sql
CREATE TABLE `test` (
  `id` int(11) unsigned NOT NULL AUTO_INCREMENT,
  `dt` datetime NOT NULL DEFAULT '1000-01-01 00:00:00',
  PRIMARY KEY (`id`),
  KEY `idx_dt` (`dt`)
) ENGINE=InnoDB AUTO_INCREMENT=2 DEFAULT CHARSET=utf8mb4;

insert into `test`(`dt`) values(`2020-05-14 12:37:06`);
```

interpolateParams=true
```golang
package main

import (
	"database/sql"
	"fmt"
	"testing"
	"time"

	_ "github.com/go-sql-driver/mysql"
)

var Loc = time.FixedZone("Asia/Shanghai", 8*60*60)

var DB *sql.DB

func init() {
	dsn := fmt.Sprintf("%s:%s@tcp(%s)/%s", "root", "123456", "127.0.0.1", "chanxuehong")
	dsn += "?collation=utf8mb4_bin&clientFoundRows=false&interpolateParams=false&loc=Asia%2FShanghai&maxAllowedPacket=0&multiStatements=false&parseTime=true&timeout=5000ms&time_zone=%27Asia%2FShanghai%27"
	db, err := sql.Open("mysql", dsn)
	if err != nil {
		fmt.Println(err)
		return
	}
	if err = db.Ping(); err != nil {
		fmt.Println(err)
		return
	}
	db.SetMaxOpenConns(10)
	db.SetMaxIdleConns(5)
	db.SetConnMaxLifetime(30 * time.Minute)

	DB = db
}

func Benchmark1(b *testing.B) {
	t := time.Date(2020, 05, 14, 12, 37, 06, 0, Loc)
	var id int

	stmt, err := DB.Prepare("select id from `test` where `dt`=?")
	if err != nil {
		b.Error(err)
		return
	}
	defer stmt.Close()

	b.ReportAllocs()
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		stmt.QueryRow("select id from `test` where `dt`=?", t).Scan(&id)
	}
}
```

```shell
➜  test1 go test -bench=. -benchtime=1m
goos: darwin
goarch: amd64
pkg: test1
Benchmark1-8   	32202482	      2263 ns/op	     224 B/op	       8 allocs/op
PASS
ok  	test1	75.836s
➜  test1 go test -bench=. -benchtime=1m
goos: darwin
goarch: amd64
pkg: test1
Benchmark1-8   	32203111	      2318 ns/op	     224 B/op	       8 allocs/op
PASS
ok  	test1	77.265s
➜  test1 go test -bench=. -benchtime=1m
goos: darwin
goarch: amd64
pkg: test1
Benchmark1-8   	32117144	      2360 ns/op	     224 B/op	       8 allocs/op
PASS
ok  	test1	78.529s
➜  test1
➜  test1
➜  test1
➜  test1
➜  test1
➜  test1
➜  test1
➜  test1 go test -bench=. -benchtime=1m
goos: darwin
goarch: amd64
pkg: test1
Benchmark1-8   	31963071	      2259 ns/op	     224 B/op	       8 allocs/op
PASS
ok  	test1	75.172s
➜  test1 go test -bench=. -benchtime=1m
goos: darwin
goarch: amd64
pkg: test1
Benchmark1-8   	32266390	      2259 ns/op	     224 B/op	       8 allocs/op
PASS
ok  	test1	75.459s
➜  test1 go test -bench=. -benchtime=1m
goos: darwin
goarch: amd64
pkg: test1
Benchmark1-8   	32220984	      2269 ns/op	     224 B/op	       8 allocs/op
PASS
ok  	test1	75.653s
➜  test1
```