@dolmen 2020-06-01
As the aim of this PR is speed, it should keep one feature of the original implementation : appending the formatted time to an existing `[]byte` to avoid temporary `[]byte` allocation and copy.

So I think that `formatDateTime(t time.Time) (buf [32]byte, n int, err error)` should be replaced by a `appendDateTime(buf []byte, t time.Time) (n int, err error)`.